+++ api/V0ShowAllTasks.go
@@ -1,8 +1,12 @@
 package api
 
 import (
+	"encoding/json"
 	"net/http"
 
+	mesosutil "github.com/AVENTER-UG/mesos-util"
+	mesosproto "github.com/AVENTER-UG/mesos-util/proto"
+	util "github.com/AVENTER-UG/util/util"
 	"github.com/sirupsen/logrus"
 )
 
@@ -16,11 +20,30 @@ func (e *API) V0ShowAllTasks(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	logrus.Debug("HTTP GET V0ShowAllTasks")
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	w.Header().Set("Api-Service", "v0")
+
+	logrus.WithField("func", "api.V0ShowAllTasks").Debug("HTTP GET V0ShowAllTasks")
 
 	keys := e.Redis.GetAllRedisKeys(e.Framework.FrameworkName + ":*")
 
+	var list []mesosutil.Command
+
 	for keys.Next(e.Redis.CTX) {
-		logrus.Info("keys: ", keys.Val())
+		// ignore redis keys if they are not mesos tasks
+		if e.Redis.CheckIfNotTask(keys) {
+			continue
+		}
+
+		key := e.Redis.GetRedisKey(keys.Val())
+		task := mesosutil.DecodeTask(key)
+
+		task.Environment = mesosproto.Environment{}
+
+		list = append(list, task)
 	}
+
+	out, _ := json.Marshal(&list)
+	d := e.ErrorMessage(0, "V0ShowAllTasks", util.PrettyJSON(out))
+	w.Write(d)
 }