+++ changelog.md
@@ -22,6 +22,7 @@
 - ADD: Mesos CLI restart and update service.
 - ADD: Support for `node.platform.os` and `node.platform.arch` constraint
 - ADD: Support of docker-compose command [restart](https://docs.docker.com/compose/compose-file/#read_only) 
+- ADD: Show all Tasks as API call and mesos-cli command.
   
 ## v0.4.0